requirements.txt
pydantic==2.11.7
google-generativeai==0.8.5
dateparser==1.2.2
python-docx==1.2.0
pymupdf==1.24.9
pyyaml==6.0.2
regex==2025.7.34
python-dotenv==1.0.1
numpy
pandas
tqdm
opencv-python
